second commit

+++ controllers/org.js
@@ -1,7 +1,6 @@
 const path = require("path");
 const fs = require("fs/promises");
 
-
 //////////////////////// read file ///////////////////////////////
 const read = async (req, res) => {
   try {
@@ -137,52 +136,68 @@ const deleteUser = async (req, res) => {
   }
 };
 
-
 /////////////////////////////////////////User upload///////////////////////////////
 
 const upload = async (req, res) => {
+  try {
+    const fsPath = path.join(process.cwd(), "./db.json");
+    const openFile = await fs.open(fsPath, "r+");
+    const data = await fs.readFile(openFile, { encoding: "utf8" });
+    const json = JSON.parse(data);
+    const id = req.params.id;
 
-    try {
-      const fsPath = path.join(process.cwd(), "./db.json");
-      const openFile = await fs.open(fsPath, "r+");
-      const data = await fs.readFile(openFile, { encoding: "utf8" });
-      const json = JSON.parse(data);
-      const id = req.params.id;
-         
-      // Find the item to upload
-      const itemIndex = json.findIndex((item) => item.id == req.params.id);
-      
-   // if given user doesn't exists the rm remove the uploaded image
-      if (itemIndex === -1) {
-           await fs.rm(path.join(process.cwd(), "public/profile_pic"));
-           return res.status(404).json({description: "User not found"});
-      }
- 
-       // if user have already set the profilePic then remove the previous and upload the new pic logic 
-       if(json[itemIndex].profile_pic){
-         
-            await fs.rm(path.join(process.cwd(), json[itemIndex].profile_pic))
-       }
-          
-       const updatedOrg = {...json[itemIndex], profile_pic: path.join('public', 'profile_pic', req.file.filename)};
-       json.splice(itemIndex, 1 , updatedOrg);
+    // check extension
 
-      
-        
-      // Write the updated data back to the file
-      await fs.writeFile(fsPath, JSON.stringify(json));
-  
-      res.json(updatedOrg);
-
-
-    } catch (error) {
-      console.log(error);
-      res.status(500).send("Server error");
+    if (!req.file) {
+      return res
+        .status(400)
+        .json({
+          description:
+            "please give an appropriate image formate.eg ['.jpg','.png', '.gif', '.webp', '.bmp']",
+        });
     }
 
+    // Find the item to upload
+    const itemIndex = json.findIndex((item) => item.id == req.params.id);
 
-        
+    // if given user doesn't exists the rm remove the uploaded image
+    if (itemIndex === -1) {
+      await fs.rm(path.join(process.cwd(), "public/profile_pic"));
+      return res.status(404).json({ description: "User not found" });
+    }
 
-  };
+    // if user have already set the profilePic then remove the previous and upload the new pic logic
+
+     
+    try {
+
+      if (json[itemIndex].profile_pic) {
+        await fs.rm(path.join(process.cwd(), json[itemIndex].profile_pic));
+      }
+      
+    } catch (error) {
+
+      if(!(error.code === "ENOENT")) {
+        throw error;
+      }
+      
+    }
+    
+
+    const updatedOrg = {
+      ...json[itemIndex],
+      profile_pic: path.join("public", "profile_pic", req.file.filename),
+    };
+    json.splice(itemIndex, 1, updatedOrg);
+
+    // Write the updated data back to the file
+    await fs.writeFile(fsPath, JSON.stringify(json));
+
+    res.json(updatedOrg);
+  } catch (error) {
+    console.log(error);
+    res.status(500).send("Server error");
+  }
+};
 
 module.exports = { read, create, update, searchUerById, deleteUser, upload };
